+++ src/components/modal/modal.module.css
@@ -7,27 +7,45 @@
     flex-direction: column;
     justify-content: center;
     align-items: center;
-    gap: 10px;
+    gap: 20px;
+    width: auto;
+}
+
+.container:focus-visible {
+    outline: none;
 }
 
 .title {
-    width: 100%;
+    text-wrap: nowrap;
     text-align: center;
     color: pink;
     text-shadow: 2px 0 red, -2px 0 red, 0 2px red, 0 -2px red,
         1px 1px red, -1px -1px red, 1px -1px red, -1px 1px red;
 }
 
+.name {
+    text-align: center;
+    color: pink;
+    text-shadow: 2px 0 red, -2px 0 red, 0 2px red, 0 -2px red,
+        1px 1px red, -1px -1px red, 1px -1px red, -1px 1px red;
+    text-wrap: balance;
+}
+
 .image {
-    width: 60vh;
-    height: 60vh;
-    aspect-ratio: 1;
-    object-fit: cover;
-    mask:
-        radial-gradient(at 70% 31%, #000 29%, #0000 30%),
-        radial-gradient(at 30% 31%, #000 29%, #0000 30%),
-        linear-gradient(#000 0 0) bottom/100% 50% no-repeat;
-    clip-path: polygon(-41% 0, 50% 91%, 141% 0);
+    width: 56vh;
+    height: 56vh;
+    background-size: cover;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    background-position: center;
+}
+
+.frame {
+    width: 58vh;
+    height: 58vh;
+    border-radius: 30px;
+    box-shadow: 1px 0px 60px red;
 }
 
 .background {
@@ -35,6 +53,11 @@
     background-size: 20vh;
     background-position: bottom;
     animation: appear 2s ease-in-out;
+}
+
+.confetti {
+    position: absolute;
+    top: 0;
 }
 
 @keyframes appear {
@@ -48,8 +71,13 @@
 }
 
 @media (max-width: 600px) {
+    .frame {
+        width: 90vw;
+        height: 90vw;
+    }
+
     .image {
-        width: 40vh;
-        height: 40vh;
+        width: 88vw;
+        height: 88vw;
     }
 }
