+++ src/utils/types.ts
@@ -1,11 +1,11 @@
 export interface IBreed {
-    weight: { metric: string };
-    name: string;
-    temperament?: string;
-  }
-  
+  weight: { metric: string };
+  name: string;
+  temperament?: string;
+}
+
 export interface IDog {
-    breeds: IBreed[];
-    id: string;
-    url: string;
+  breeds: IBreed[];
+  id: string;
+  url: string;
 }
